% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking-evaluations.R
\name{block.ids.from.blocking}
\alias{block.ids.from.blocking}
\title{Returns the block ids associated with a blocking method.}
\usage{
block.ids.from.blocking(blocking)
}
\arguments{
\item{blocking}{A list of the blocks.}
}
\value{
A list of the blocks ids that corresponds to each block
}
\description{
Returns the block ids associated with a blocking method.
}
\examples{
data("RLdata500")
klsh.blocks <- klsh(RLdata500, p=20, num.blocks=5, k=2)
block.ids.from.blocking(klsh.blocks)
}
