% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klsh.R
\name{calc_idf}
\alias{calc_idf}
\title{Function to calculate the inverse document frequency given a shingled bag of words}
\usage{
calc_idf(sack_of_bags)
}
\arguments{
\item{sack_of_bags}{Sack of bag of words}
}
\value{
Computes the inverse document frequency for a bag of words
}
\description{
Function to calculate the inverse document frequency given a shingled bag of words
}
\examples{
data(RLdata500)
data.500 <- RLdata500[-c(2,4)]
sack <- sacks_of_bags_of_words(data.500[1:3,c(-2)],k=2)
(idf <- calc_idf(sack))
match(names(sack[[1]]), names(idf))
}
