% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking-evaluations.R
\name{reduction.ratio.from.blocking}
\alias{reduction.ratio.from.blocking}
\title{Returns the reduction ratio associated with a blocking method}
\usage{
reduction.ratio.from.blocking(blocking)
}
\arguments{
\item{blocking}{The actual blocks}
}
\value{
The reduction ratio
}
\description{
Returns the reduction ratio associated with a blocking method
}
\examples{
data("RLdata500")
klsh.blocks <- klsh(RLdata500, p=20, num.blocks=5, k=2)
reduction.ratio.from.blocking(klsh.blocks)
}
