% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking-evaluations.R
\name{reduction.ratio}
\alias{reduction.ratio}
\title{Returns the reduction ratio associated with a blocking method}
\usage{
reduction.ratio(block.labels)
}
\arguments{
\item{block.labels}{A list of the blocks labels.}
}
\value{
The reduction ratio
}
\description{
Returns the reduction ratio associated with a blocking method
}
\examples{
data("RLdata500")
klsh.blocks <- klsh(RLdata500, p=20, num.blocks=5, k=2)
block.ids <- block.ids.from.blocking(klsh.blocks)
reduction.ratio(block.ids)
}
