% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klsh.R
\name{sacks_of_bags_of_words}
\alias{sacks_of_bags_of_words}
\title{Function to convert all records into a bag of tokens}
\usage{
sacks_of_bags_of_words(r.set, k, fieldwise = FALSE)
}
\arguments{
\item{r.set}{Record set}

\item{k}{Parameter k, which is the number of shingle, tokens, or grams to break the string into}

\item{fieldwise}{Flag where the defalt setting to include the record as the entire string}
}
\value{
Computes the bag of tokens for a record set
}
\description{
Function to convert all records into a bag of tokens
}
\examples{
data(RLdata500)
data.500 <- RLdata500[-c(2,4)]
sacks_of_bags_of_words(data.500[1:3,c(-2)],k=2)
}
