% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klsh.R
\name{tokenify}
\alias{tokenify}
\title{Function to token a string into its k components}
\usage{
tokenify(string, k)
}
\arguments{
\item{string}{A string or record}

\item{k}{A parameter k, which is the number of shingle, tokens, or grams to break the string into}
}
\value{
Computes the tokenized or grammed version of a string
}
\description{
Function to token a string into its k components
}
\examples{
tokenify("Alexander",2)
tokenify("Alexander Smith", 2)
}
